requirements.txt
importlib_resources

eoreader>=0.21.1
cdsapi>=0.6.1
GRSdriver >= 1.0.0

rasterio
netCDF4
numba
numpy
pandas>=2.0.3
pyproj

rioxarray>=0.15.0
s2cloudless>=1.7.1
